renders multiple spheres, offCentre intCheck broken

+++ includes/types.h
@@ -6,7 +6,7 @@
 /*   By: ccaljouw <ccaljouw@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/09/12 16:10:32 by ccaljouw      #+#    #+#                 */
-/*   Updated: 2023/09/16 14:05:29 by cariencaljo   ########   odam.nl         */
+/*   Updated: 2023/09/16 16:53:20 by cariencaljo   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,14 +31,14 @@ typedef struct ambient_s
  * @param	view_point (t_xyz) x,y,z coordinates of the camera's origing
  * @param	orientation_v (t_xyz) 3d normalized orientation vector. 
  * In range [-1,1] for each x,y,z axis
- * @param	fov	(int) Horizontal field of view in degrees in range [0,180]
+ * @param	fov	(float) Horizontal field of view in degrees in range [0,180]
  * @param	principal_axis (t_xyz) vector from camera positon to lookAt position
  */
 typedef struct camera_s
 {
 	t_xyz	view_point;
 	t_xyz	orientation_v;
-	int		fov;
+	float	fov;
 	t_xyz	look_at;
 	float	plane_dist;
 	float	c_width;
